Increment Dual
This is the next step from DualSCRdrive towards a version of Draft1 with the optos and passives. Did I make it worse or better? I've changed the numbering of the SCRs to match that in Draft1, and setup the optos as "1"+ the SCR number. e.g. SCR 3 is controlled by Opto 13. I'm not sure about the placement of R6, and not all that sure about U4. The cathode on U14 (opto that controls U4) would be driven by another power supply circuit, which would be derived from U1? But I'm not sure that's correct since U4's anode is not shared with U1. Again, did I make it worse or better? If worse, what's wrong, if better, I'll continue on.